Fix reload on configmap changes
In scenarios where the file is mounted inside a container, we won't receive any event on the config file changes, so we have to listen on the file not the parent directory. Because of a212099c, we also have to watch the directory. So, here is a fix to listen on both, but limit the rate of triggers by waiting .2 seconds before reloading, and ignoring more reloads during this .2 seconds.